extract image request/pipeline
